changefeedccl: block db-level changefeed creation for non-empty tableset
On startup, the changefeed should not be started until there exist
tables in the database to watch. If no tables exist at the time of
startup, then a tableset watcher is created and polled; the changefeed
will be started from where there is first a target table.

Fixes: #147371
Epic: CRDB-1421

Release note: none